Fix #2 rolling log file handling
It seems explicit volume map to a file locks in the pointer to that file on the container,
so I mapped to the directory and made the file name configurable.

Tested it out with log rolling every minute, it did work picking up the log entries,
but it also complained about handling the file:

```
TailFile had an error! Error: Could not read remaining bytes from old FH
    at TailFile._pollFileForChanges (/usr/src/app/node_modules/@logdna/tail-file/lib/tail-file.js:185:25) {
  meta: {
    error: "The argument 'size' is invalid. Received -209",
    code: 'ERR_INVALID_ARG_VALUE'
  }
}
log entry: 2023-10-15 17:36:56 [Server thread/WARN]: Socrates2100 moved too quickly! 14.882470404550077,-1.6211341470457228,-11.371377410519926
(node:19) Warning: Closing file descriptor 21 on garbage collection
(Use `node --trace-warnings ...` to show where the warning was created)
(node:19) [DEP0137] DeprecationWarning: Closing a FileHandle object on garbage collection is deprecated. Please close FileHandle objects explicitly using FileHandle.prototype.close(). In the future, an error will be thrown if a file descriptor is closed during garbage collection.
```

Since it was able to process subsequent entries successfully we can safely ignore
this error for now. If in the future an error is thrown instead of this warning,
it will just force the app on a restart, which would also "be good enough" :D.